created a separate class and list for storing ChangeNotifierProvider data
