+++ app/safenet/lib/main.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:safenet/core/routes.dart';
-import 'package:safenet/core/viewmodals/home_model.dart';
-import 'package:safenet/core/viewmodals/login_model.dart';
-import 'package:safenet/core/viewmodals/register_model.dart';
-import 'package:safenet/core/viewmodals/welcome_model.dart';
+import 'package:safenet/core/services/providers.dart';
+import 'package:safenet/core/services/routes.dart';
 import 'package:safenet/ui/views/welcome_view.dart';
 
 void main() {
@@ -16,15 +13,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
-      providers: [
-        ChangeNotifierProvider<WelcomeModel>(
-            create: (context) => WelcomeModel()),
-        ChangeNotifierProvider<LoginModel>(create: (context) => LoginModel()),
-        ChangeNotifierProvider<RegisterModel>(
-            create: (context) => RegisterModel()),
-        ChangeNotifierProvider<HomeViewModel>(
-            create: (context) => HomeViewModel())
-      ],
+      providers: Providers.providers,
       // TODO: create a seperate file to map all the changeNotifierProviders to their repective Models.
       child: MaterialApp(
         routes: Routes.getRoutes(),
