+++ app/safenet/lib/ui/views/user_details_entry.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:safenet/core/routes.dart';
+import 'package:safenet/core/services/routes.dart';
 import 'package:safenet/ui/constants.dart';
 import 'package:safenet/ui/customWidgets/bloodGroupDropDown.dart';
 import 'package:safenet/ui/customWidgets/customSignInButton.dart';
